Breaks down data fetching hooks, improves APIService documentation comments
