+++ src/hooks/DataFetching/PaginatedRecords.ts
@@ -1,416 +1,14 @@
-import {
-  Dispatch,
-  SetStateAction,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import { useCallback, useEffect, useMemo, useRef } from 'react';
 
-import { CANCELLED_API_REQUEST_MESSAGE } from '../constants';
-import { APIContext } from '../contexts/APIContext';
 import {
   PaginatedRequestParams,
   PaginatedResponseData,
   RecordFinderRequestController,
-  TAPIFunction,
-  TaggedAPIRequest,
-} from '../models/Utils';
-import { usePolling } from './Polling';
+} from '../../models/Utils';
+import { usePolling } from '../Polling';
+import { useAPIService } from './APIService';
+import { GetStaleWhileRevalidateFunction, QueryOptions } from './models';
 
-export interface QueryOptions {
-  loadOnMount?: boolean;
-  autoSync?: boolean;
-  revalidationKey?: string;
-}
-
-export type GetStaleWhileRevalidateFunction<T> = (data: T) => void;
-
-//#region useAPIService
-export const useAPIService = <T>(defaultValue: T, loadOnMount = false) => {
-  const baseDefaultValue = defaultValue;
-  const isComponentMountedRef = useRef(true);
-  const taggedAPIRequestsRef = useRef<TaggedAPIRequest[]>([]);
-  const { call } = useContext(APIContext);
-  const [record, setRecord] = useState<T>(defaultValue);
-  const [loaded, setLoaded] = useState(false);
-  const [loading, setLoading] = useState(loadOnMount);
-  const [busy, setBusy] = useState(false);
-  const [errorMessage, setErrorMessage] = useState('');
-
-  const load = useCallback(
-    async (apiFunction?: TAPIFunction, tag?: string, polling = false) => {
-      if (apiFunction) {
-        setErrorMessage('');
-        setLoaded(false);
-        setLoading(true);
-        const taggedAPIRequest = (() => {
-          if (tag) {
-            return {
-              id: tag,
-              loading: true,
-            } as TaggedAPIRequest;
-          }
-        })();
-        taggedAPIRequest && taggedAPIRequestsRef.current.push(taggedAPIRequest);
-        const response = await call(apiFunction)
-          .then(async (payload) => {
-            const response = await payload;
-            if (response != null) {
-              if (isComponentMountedRef.current) {
-                setRecord(response);
-              }
-            }
-            if (isComponentMountedRef.current) {
-              setLoaded(true);
-            }
-            return payload;
-          })
-          .catch((err) => {
-            if (
-              !String(err.message).match(CANCELLED_API_REQUEST_MESSAGE) &&
-              !polling
-            ) {
-              setErrorMessage(err.message);
-            }
-          });
-        if (
-          taggedAPIRequest &&
-          taggedAPIRequestsRef.current.includes(taggedAPIRequest)
-        ) {
-          taggedAPIRequestsRef.current.splice(
-            taggedAPIRequestsRef.current.indexOf(taggedAPIRequest),
-            1
-          );
-        }
-        if (isComponentMountedRef.current) {
-          setLoading(false);
-        }
-        return response;
-      }
-    },
-    [call]
-  );
-
-  const resetRef = useRef(() => {
-    setRecord(baseDefaultValue);
-    setLoaded(false);
-    setLoading(loadOnMount);
-    setBusy(false);
-    setErrorMessage('');
-  });
-
-  useEffect(() => {
-    isComponentMountedRef.current = true;
-    return () => {
-      isComponentMountedRef.current = false;
-    };
-  }, []);
-
-  return {
-    errorMessage,
-    setErrorMessage,
-    load,
-    loaded,
-    loading,
-    record,
-    setLoaded,
-    setRecord,
-    busy,
-    setBusy,
-    taggedAPIRequests: taggedAPIRequestsRef.current,
-    reset: resetRef.current,
-  };
-};
-//#endregion
-
-type Awaited<T> = T extends PromiseLike<infer U> ? U : T;
-
-//#region useMutation
-export const useMutation = <MutateFunction extends TAPIFunction>(
-  inputMutate: MutateFunction
-) => {
-  const {
-    load: baseMutate,
-    loading: mutating,
-    record: mutatedRecord,
-    loaded: mutated,
-    setLoaded: setMutated,
-    ...rest
-  } = useAPIService<Awaited<ReturnType<MutateFunction>> | null>(null);
-
-  const inputMutateRef = useRef(inputMutate);
-  inputMutateRef.current = inputMutate;
-
-  const mutate = useCallback(
-    (...args: any) => {
-      return baseMutate(() => inputMutateRef.current(...args));
-    },
-    [baseMutate]
-  ) as MutateFunction;
-
-  return {
-    mutate,
-    mutating,
-    mutated,
-    setMutated,
-    mutatedRecord,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region useCreate
-export const useCreate = <CreateFunction extends TAPIFunction>(
-  inputCreate: CreateFunction
-) => {
-  const {
-    mutate: create,
-    mutating: creating,
-    mutated: created,
-    setMutated: setCreated,
-    mutatedRecord: createdRecord,
-    ...rest
-  } = useMutation(inputCreate);
-
-  return {
-    create,
-    creating,
-    created,
-    setCreated,
-    createdRecord,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region useUpdate
-export const useUpdate = <UpdateFunction extends TAPIFunction>(
-  inputUpdate: UpdateFunction
-) => {
-  const { create, creating, created, createdRecord, setCreated, ...rest } =
-    useCreate(inputUpdate);
-
-  return {
-    update: create,
-    updating: creating,
-    updated: created,
-    updatedRecord: createdRecord,
-    setUpdated: setCreated,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region useDelete
-export const useDelete = <DeleteFunction extends TAPIFunction>(
-  inputDelete: DeleteFunction
-) => {
-  const {
-    mutate: _delete,
-    mutating: deleting,
-    mutated: deleted,
-    setMutated: setDeleted,
-    mutatedRecord: deletedRecord,
-    ...rest
-  } = useMutation(inputDelete);
-
-  return {
-    _delete,
-    deleting,
-    deleted,
-    setDeleted,
-    deletedRecord,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region useRecord
-export interface UseRecordOptions<LoadableRecord> extends QueryOptions {
-  defaultValue?: LoadableRecord;
-}
-export const useRecord = <LoadableRecord>(
-  recordFinder?: TAPIFunction<LoadableRecord>,
-  {
-    defaultValue,
-    loadOnMount = true,
-    revalidationKey,
-  }: UseRecordOptions<LoadableRecord> = {}
-) => {
-  // Refs
-  const isInitialMountRef = useRef(true);
-  const recordFinderRef = useRef(recordFinder);
-  recordFinderRef.current = recordFinder;
-
-  const {
-    load: baseLoad,
-    loading,
-    errorMessage,
-    busy,
-    record,
-    setRecord,
-    ...rest
-  } = useAPIService(defaultValue, loadOnMount);
-
-  const load = useCallback(
-    (...args: any) => {
-      return baseLoad(() => {
-        if (recordFinderRef.current) {
-          return recordFinderRef.current(...args);
-        }
-        return Promise.resolve();
-      });
-    },
-    [baseLoad]
-  ) as NonNullable<typeof recordFinder>;
-
-  useEffect(() => {
-    if (loadOnMount && isInitialMountRef.current) {
-      load();
-    }
-  }, [load, loadOnMount]);
-
-  useEffect(() => {
-    if (!isInitialMountRef.current && revalidationKey) {
-      load();
-    }
-  }, [load, revalidationKey]);
-
-  useEffect(() => {
-    isInitialMountRef.current = false;
-    return () => {
-      isInitialMountRef.current = true;
-    };
-  }, []);
-
-  return {
-    load,
-    loading,
-    errorMessage,
-    record: record ?? null,
-    setRecord: setRecord as Dispatch<SetStateAction<LoadableRecord | null>>,
-    busy,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region useRecords
-export const useRecords = <LoadableRecord>(
-  recordFinder?: TAPIFunction<LoadableRecord[]>,
-  { ...inputRest }: QueryOptions = {}
-) => {
-  const { record, setRecord, ...rest } = useRecord(recordFinder, {
-    defaultValue: [],
-    ...inputRest,
-  });
-
-  return {
-    records: record!,
-    setRecords: setRecord as Dispatch<SetStateAction<LoadableRecord[]>>,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region useCacheableData
-export type CacheableDataFinderOptions = {
-  getRequestController?: (controller: RecordFinderRequestController) => void;
-  getStaleWhileRevalidate?: GetStaleWhileRevalidateFunction<any>;
-};
-export type CacheableDataFinder<Data> = (
-  options: CacheableDataFinderOptions
-) => Promise<Data>;
-export interface CacheableDataOptions<Data> extends QueryOptions {
-  defaultValue?: Data;
-}
-export const useCacheableData = <Data>(
-  recordFinder?: CacheableDataFinder<Data>,
-  {
-    defaultValue,
-    loadOnMount = true,
-    revalidationKey,
-  }: CacheableDataOptions<Data> = {}
-) => {
-  //#region Refs
-  const isInitialMountRef = useRef(true);
-
-  const recordFinderRef = useRef(recordFinder);
-  recordFinderRef.current = recordFinder;
-
-  const pendingDataRequestControllerRef =
-    useRef<RecordFinderRequestController | null>(null);
-  //#endregion
-
-  const {
-    load: baseLoad,
-    loading,
-    errorMessage,
-    record,
-    setRecord,
-    ...rest
-  } = useAPIService(defaultValue, loadOnMount);
-
-  const load = useCallback(() => {
-    return baseLoad(async () => {
-      if (recordFinderRef.current) {
-        if (pendingDataRequestControllerRef.current) {
-          pendingDataRequestControllerRef.current.cancelRequest();
-          pendingDataRequestControllerRef.current = null;
-        }
-
-        const responseData = await recordFinderRef
-          .current({
-            getRequestController: (requestController) => {
-              pendingDataRequestControllerRef.current = requestController;
-            },
-            getStaleWhileRevalidate: (data) => {
-              setRecord(data);
-            },
-          })
-          .finally(() => {
-            pendingDataRequestControllerRef.current = null;
-          });
-
-        return responseData;
-      }
-    });
-  }, [baseLoad, setRecord]);
-
-  useEffect(() => {
-    if (loadOnMount && isInitialMountRef.current) {
-      load();
-    }
-  }, [load, loadOnMount]);
-
-  useEffect(() => {
-    if (!isInitialMountRef.current && revalidationKey) {
-      load();
-    }
-  }, [load, revalidationKey]);
-
-  useEffect(() => {
-    isInitialMountRef.current = false;
-    return () => {
-      isInitialMountRef.current = true;
-    };
-  }, []);
-
-  return {
-    load,
-    loading,
-    errorMessage,
-    data: record,
-    setData: setRecord,
-    ...rest,
-  };
-};
-//#endregion
-
-//#region usePaginatedRecords
 export type PaginatedRecordsFinderOptions<
   PaginatedResponseDataExtensions extends Record<string, any> = any
 > = PaginatedRequestParams & {
@@ -467,7 +65,7 @@ export const usePaginatedRecords = <
     refreshInterval,
   }: PaginatedRecordsOptions<DataRow> = {}
 ) => {
-  // Refs
+  //#region Ref
   const isInitialMountRef = useRef(true);
   const pendingRecordRequestControllers = useRef<
     RecordFinderRequestController[]
@@ -489,6 +87,7 @@ export const usePaginatedRecords = <
   const lastLoadedPageRef = useRef<
     ResponsePage<DataRow, PaginatedResponseDataExtensions> | undefined
   >(undefined);
+  //#endregion
 
   const {
     load: loadFromAPIService,
@@ -723,4 +322,3 @@ export const usePaginatedRecords = <
     ...rest,
   };
 };
-//#endregion
