+++ src/hooks/DataFetching/models.ts
@@ -1,0 +1,7 @@
+export interface QueryOptions {
+  loadOnMount?: boolean;
+  autoSync?: boolean;
+  revalidationKey?: string;
+}
+
+export type GetStaleWhileRevalidateFunction<Data> = (data: Data) => void;
